Move resorce loading from view init() to AppConfig ready() method
This is done to load the resources only once on startup instead of loading it everytime the view is initialized which is done whenever a request is made to the resources endpoint.
